backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
httpx
# for sqlite
aiosqlite
openai
pypdf2